sections: only display after section start
fixes #28
It was wrong to give all the contents to the preview/display command and
then search on the output, because the command could change the format
completely.
So now we submit to the converter program only data after the section
start, and we don't need to search the output anymore.